requirements.txt
discord.py
requests
matplotlib
python-dotenv
